requirements.txt
openai
azure-identity
streamlit